refactor(websocket): change package name from entity to handler
