+++ socket/src/main/scala/io/github/positionpal/server/ws/WebSocketHandlers.scala
@@ -6,7 +6,7 @@ import akka.stream.scaladsl.{Flow, Sink, Source}
 import akka.stream.typed.scaladsl.{ActorSink, ActorSource}
 import akka.http.scaladsl.model.ws.{Message, TextMessage}
 import akka.stream.{CompletionStrategy, OverflowStrategy}
-import io.github.positionpal.entity.Handler.{Commands, WebSocketHandler}
+import io.github.positionpal.handler.Handler.{Commands, WebSocketHandler}
 import io.github.positionpal.message.ChatMessageADT.ChatMessage
 
 /**
@@ -21,13 +21,13 @@ object WebSocketHandlers:
    */
   def websocketHandler(incomingActorRef: ActorRef[Commands]): Flow[Message, Message, WebSocketHandler] =
 
-    import io.github.positionpal.entity.Handler.Commands.{
+    import io.github.positionpal.handler.Handler.Commands.{
       IncomingMessage,
       OutgoingMessage,
       StreamCompletedSuccessfully,
       StreamCompletedWithException
     }
-    import io.github.positionpal.entity.Handler.{Commands, WebSocketHandler}
+    import io.github.positionpal.handler.Handler.{Commands, WebSocketHandler}
 
     val incomingMessage: Sink[Message, NotUsed] =
       Flow[Message].map:
